ai-service/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.2
requests==2.33.1
gunicorn==21.2.0